# [CANN]Opt ROPE optimization