### What does this PR do?
This PR optimizes the `ROPE` operator to improve inference performance. The performance improvement is **approximately 10%**. For detailed performance gains, please refer to the test report.

**Optimizations include**:
1. Reducing unnecessary memory allocations and deallocations.
2. Eliminating redundant transpose operations by optimizing matrix shapes.

### Test
**Model:** Qwen2-0.5B
```
# before:
system
You are a helpful assistant.
user
Building a website can be done in 10 steps:
assistant
That sounds like a lot of work! But here are 10 steps you can take to build a website:

  1. Identify the purpose of your website: Before you start building your website, you need to know what you want to achieve with it. Do you want to sell products, share information, or promote your business?
  2. Choose a domain name: You will need a website domain name. There are many websites that allow you to create your own domain name, or you can choose a domain name from a list that is available on the web.
  3. Choose a hosting provider: There are many hosting providers available, and you need to choose one that suits your needs. Make sure you choose a hosting provider that provides reliable and fast hosting services.
  4. Choose a website builder: There are many websites that offer website builders, and you can choose one that suits your needs. Make sure you choose a website builder that provides user-friendly interfaces and easy-to-use features.
  5. Choose a domain name: You will need a website domain name. Make sure you choose a domain name that is easy to remember and relevant to your business.
  6. Choose a hosting provider: You will need a website host provider. Make sure you choose a hosting provider that provides reliable and fast hosting services.
  7. Choose a website builder: You will need a website builder. Make sure you choose a website builder that provides user-friendly interfaces and easy-to-use features.
  8. Choose a domain name: You will need a website domain name. Make sure you choose a domain name that is easy to remember and relevant to your business.
  9. Choose a hosting provider: You will need a website host provider. Make sure you choose a hosting provider that provides reliable and fast hosting services.
  10. Choose a website builder: You will need a website builder. Make sure you choose a website builder that provides user-friendly interfaces and easy-to-use features.

> 
llama_perf_sampler_print:    sampling time =     752.70 ms /   439 runs   (    1.71 ms per token,   583.23 tokens per second)
llama_perf_context_print:        load time =    4418.41 ms
llama_perf_context_print: prompt eval time =      23.22 ms /    31 tokens (    0.75 ms per token,  1335.17 tokens per second)
llama_perf_context_print:        eval time =    8157.39 ms /   407 runs   (   20.04 ms per token,    49.89 tokens per second)
llama_perf_context_print:       total time =   10539.05 ms /   438 tokens

# After
system
You are a helpful assistant.
user
Building a website can be done in 10 steps:
assistant
Building a website can be done in 10 steps:

1. Decide what you want your website to look like. This will depend on the type of business or product you are building, as well as the target audience. 

2. Choose a platform to build your website on. There are many options available, including WordPress, Wix, Squarespace, and more. Choose one that suits your needs and budget.

3. Select a design and layout for your website. You can use a website builder or create your own design using a software tool. Make sure that your website is easy to navigate and visually appealing.

4. Choose a domain name and registrar. This will make your website accessible to users who are using a specific domain name.

5. Choose a hosting provider. This will ensure that your website is available to users who are using a specific hosting provider.

6. Design your website. You can use a website builder or create your own design using a software tool. Make sure that your website is easy to navigate and visually appealing.

7. Publish your website. Once you have designed your website and chosen a hosting provider, you can publish it on the internet.

8. Test your website. Make sure that it is functional and accessible to users.

9. Optimize your website. This will help increase your website's traffic and search engine rankings.

10. Launch your website. Once your website is published and optimized, you can start promoting it on social media and other marketing channels to increase traffic and sales.

Note that building a website can be a complex process and may require a lot of time and effort. It is recommended that you start by building a basic website and then build it up step by step.

> 
llama_perf_sampler_print:    sampling time =     652.65 ms /   378 runs   (    1.73 ms per token,   579.18 tokens per second)
llama_perf_context_print:        load time =    4414.03 ms
llama_perf_context_print: prompt eval time =      21.73 ms /    31 tokens (    0.70 ms per token,  1426.80 tokens per second)
llama_perf_context_print:        eval time =    6267.05 ms /   346 runs   (   18.11 ms per token,    55.21 tokens per second)
llama_perf_context_print:       total time =   11959.63 ms /   377 tokens
```